refactor: 은행 선택 UI 개선
- BankType 대신 BankUiModel 사용하도록 변경
- 카드 클릭 이벤트 핸들러 추가
