+++ app/src/androidTest/java/woowacourse/payments/BankTest.kt
@@ -6,10 +6,11 @@ import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.payments.domain.BankType
+import woowacourse.payments.ui.model.BankUiModel
+import woowacourse.payments.ui.model.toLocalBankUiModel
 import woowacourse.payments.ui.newcard.banks.Bank
 import woowacourse.payments.ui.newcard.banks.BanksTestTag
 
@@ -19,22 +20,22 @@ class BankTest {
 
     @Test
     fun `은행_컨테이너를_클릭하면_해당_은행타입으로_onSelectCard가_호출된다`() {
-        var bankType: BankType? = null
+        var bankUiModel: BankUiModel? = null
         composeTestRule.setContent {
-            Bank(BankType.BC, { bankType = it })
+            Bank(BankType.BC.toLocalBankUiModel(), { bankUiModel = it })
         }
         this@BankTest.composeTestRule
             .onNodeWithTag(BanksTestTag.BANK_CONTAINER_TAG)
             .assertIsEnabled()
             .performClick()
 
-        assert(bankType == BankType.BC)
+        assert(bankUiModel == BankType.BC.toLocalBankUiModel())
     }
 
     @Test
     fun `은행_이름이_UI에_표시된다`() {
         composeTestRule.setContent {
-            Bank(BankType.BC, {})
+            Bank(BankType.BC.toLocalBankUiModel(), {})
         }
         composeTestRule
             .onNodeWithTag(BanksTestTag.BANK_NAME_TAG, useUnmergedTree = true)
