+++ app/src/main/java/woowacourse/payments/ui/debug/fixture/Fixture.kt
@@ -10,6 +10,8 @@ val paymentCardUiModelSample =
         "1234567812345678",
         "1225",
         "빰".repeat(30),
+        "1234",
+        1
     )
 
 val paymentCardUiModelSamples =
