+++ app/src/main/java/woowacourse/payments/ui/newcard/banks/Bank.kt
@@ -18,20 +18,20 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.domain.BankType
 import woowacourse.payments.ui.cards.core.mapper.asPainter
+import woowacourse.payments.ui.model.BankUiModel
 import woowacourse.payments.ui.model.toLocalBankUiModel
 import woowacourse.payments.ui.newcard.banks.BanksTestTag.BANK_CONTAINER_TAG
 
 @Composable
 fun Bank(
-    bankType: BankType,
-    onSelectedCard: (BankType) -> Unit,
+    bankUiModel: BankUiModel,
+    onSelectedCard: (BankUiModel) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val bankUiModel = bankType.toLocalBankUiModel() ?: return
     Column(
         modifier =
             modifier
-                .clickable { onSelectedCard(bankType) }
+                .clickable { onSelectedCard(bankUiModel) }
                 .testTag(BANK_CONTAINER_TAG),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
@@ -54,5 +54,5 @@ fun Bank(
 @Preview
 @Composable
 fun BankPreview() {
-    Bank(BankType.BC, {})
+    Bank(BankType.BC.toLocalBankUiModel(), {})
 }
