+++ app/src/main/java/woowacourse/payments/ui/newcard/banks/BanksBottomSheet.kt
@@ -1,7 +1,5 @@
 package woowacourse.payments.ui.newcard.banks
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -11,17 +9,17 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.domain.BankType
+import woowacourse.payments.ui.model.BankUiModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun BanksBottomSheet(
     sheetState: SheetState,
     onDismissRequest: () -> Unit,
-    onSelectCard: (BankType) -> Unit,
+    onSelectCard: (BankUiModel) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     ModalBottomSheet(
